Safari-609.1.11.1/LayoutTests/fast/history/page-cache-running-audiocontext-expected.txt

Tests that a page with a running AudioContext goes into the back/forward cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the back/forward cache
PASS successfullyParsed is true

TEST COMPLETE
